Extract terminal and diff panel view logic
